Created pkg and moved files for authconfigmap. Renamed functions
